+++ lib/linzer.rb
@@ -70,5 +70,9 @@ module Linzer
       verify_key = block_given? ? (yield keyid) : key
       Linzer.verify(verify_key, message, signature, no_older_than: no_older_than)
     end
+
+    def signature_base(message, components, parameters)
+      Linzer::Common.signature_base(message, components, parameters)
+    end
   end
 end